+++ src/lib/get-repo-url.ts
@@ -1,3 +1,4 @@
+import axios from 'axios'
 import * as path from 'node:path'
 
 import { getRepoUrl } from './helpers/git-utils'
@@ -17,4 +18,21 @@ export async function getRepoBaseName() {
   const repoName = parsed.name.trim()
 
   return repoName
+}
+
+export async function listNpmPackages() {
+  try {
+    const org = '@ax-sh'
+    const url = `https://api.github.com/orgs/${org}/packages?package_type=npm`
+    const response = await axios.get(url, {
+      headers: {
+        Authorization: `token ${process.env.NPM_TOKEN}`,
+        // Authorization: `token ${process.env.GITHUB_TOKEN}`,
+        Accept: 'application/vnd.github.v3+json',
+      },
+    })
+    console.log(response.data)
+  } catch (error) {
+    console.error('Error fetching packages:', error.response.data)
+  }
 }