+++ src/lib/helpers/vite-config-parser.ts
@@ -45,6 +45,21 @@ export function getViteConfigPlugins(sourceFile: SourceFile) {
     .getInitializerIfKindOrThrow(ts.SyntaxKind.ArrayLiteralExpression)
 }
 
+export function getViteConfigServer(sourceFile: SourceFile) {
+  const configObject = getViteDefineConfigCallOptions(sourceFile)
+  let serverProp = configObject.getProperty('server')
+  if (!serverProp) {
+    // Create the server property if it doesn't exist
+    serverProp = configObject.addPropertyAssignment({
+      name: 'server',
+      initializer: '{}',
+    })
+  }
+  return serverProp
+    .asKindOrThrow(ts.SyntaxKind.PropertyAssignment)
+    .getInitializerIfKindOrThrow(ts.SyntaxKind.ObjectLiteralExpression)
+}
+
 export function getVitestConfigTest(sourceFile: SourceFile) {
   const configObject = getViteDefineConfigCallOptions(sourceFile)
   return configObject
@@ -80,12 +95,23 @@ export function addVitestDepsForReact(sourceFile: SourceFile) {
     //   include: ['src/**'],
     // },
   }
+
   const testObjLiteral = getVitestConfigTest(sourceFile)
   for (const [name, value] of Object.entries(config)) {
-    testObjLiteral.insertPropertyAssignment(0, {
-      name,
-      initializer: `'${value}'`,
-    })
+    try {
+      const existingProperty = testObjLiteral.getPropertyOrThrow(name)
+      if (existingProperty) {
+        // for now replacing environment
+        // Replace the existing property with the new one
+        existingProperty.replaceWithText(`${name}: '${value}'`)
+      }
+    } catch (e) {
+      console.log('Error:', e, name, 'does not exist adding')
+      testObjLiteral.insertPropertyAssignment(0, {
+        name,
+        initializer: `'${value}'`,
+      })
+    }
   }
   return sourceFile
 }